requirements.txt
# Dependências para o bot de exemplo e Worker
requests>=2.31.0
botcity-framework-web